code restructured | code for platform authorizer and platform processes is separated
