--- build.sh
@@ -1,2 +1,0 @@
-rm daemon_authenticate.zip
-zip -r daemon_authenticate.zip * -x \*venv\*
