+++ constant.py
@@ -1,0 +1,11 @@
+NETWORKS = {
+    999: {
+        'db': {'DB_HOST': '',
+               'DB_USER': '',
+               'DB_PASSWORD': '',
+               'DB_NAME': '',
+               'DB_PORT': 3306
+               }
+    }
+}
+METRICS_NETWORK_ID  = 999
