+++ platform-usage-authorizer/build.sh
@@ -1,0 +1,2 @@
+rm daemon_authorizer.zip
+zip -r daemon_authorizer.zip * -x \*venv\*
