+++ platform-usage-process/build.sh
@@ -1,0 +1,2 @@
+rm platform_usage_process.zip
+zip -r platform_usage_process.zip * -x \*venv\*
