+++ platform-usage-process/requirement.txt
@@ -1,0 +1,1 @@
+pymysql==0.9.2
